requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
faiss-cpu
gunicorn
pillow
